data extract fix
