+++ flows/data_extraction.py
@@ -1,93 +1,73 @@
-import io
 import os
-import uuid
+import random
+import re
 from pathlib import Path
 
 import requests
 from bs4 import BeautifulSoup
-from dotenv import load_dotenv
 from prefect import flow, get_run_logger, task
-from prefect_aws import S3Bucket
-from utils import get_base_url
-
-load_dotenv()
-
-block_s3_bucket = os.getenv("PREFECT_BLOCK_S3_BUCKET")
-s3_bucket = S3Bucket.load(block_s3_bucket)
 
 
-@task
-def fetch_urls_from_s3(file: str = "data/source.txt") -> list[str]:
-    """Reads the content of a file from S3 and returns it as a list of strings."""
-    logger = get_run_logger()
-    logger.info(f"Reading file from S3: {file}")
-
-    try:
-        content = s3_bucket.read_path(file)
-        logger.info(f"Successfully read {len(content)} bytes from S3.")
-        content_str = content.decode("utf-8")
-        return (
-            content_str.splitlines()
-        )  # Return a list of strings, each representing a line
-    except Exception as e:
-        logger.error(f"Error reading file from S3: {e}")
-        return []
-
-
-@task
-def put_data_to_s3(data: bytes, s3_key: str) -> None:
-    logger = get_run_logger()
-    logger.info(f"Uploading data to S3: {s3_key}")
-    file_obj = io.BytesIO(data)
-    try:
-        s3_bucket.upload_from_file_object(file_obj, s3_key)
-        logger.info(f"Successfully uploaded data to s3://{block_s3_bucket}/{s3_key}")
-    except Exception as e:
-        logger.error(f"Error uploading to S3: {e}")
+def get_base_url(url):
+    # Extract the base url: protocol, subdomain, and domain
+    base_re = r"^.+?[^\/:](?=[?\/]|$)"
+    return re.findall(base_re, url)[0]
 
 
 @task
 def get_pdf_urls(url):
+    # Extract the base url: protocol, subdomain, and domain
     base_url = get_base_url(url)
-    logger = get_run_logger()
 
+    # Stop processing if the request wasn't successful
     try:
         r = requests.get(url)
-        r.raise_for_status()
-    except requests.RequestException as e:
-        logger.error(f"Error fetching PDF URLs from {url}: {e}")
+    except requests.RequestException:
+        logger = get_run_logger()
+        logger.info("Error fetching PDF URLs... " + url)
         return set()
 
-    bs = BeautifulSoup(r.content, "html.parser")
+    if r.status_code != 200:
+        return set()
+
+    # Extract all a-tags from the website HTML
+    bs = BeautifulSoup(r.content)
     a_tags = bs.findAll("a")
 
-    hrefs = [
-        tag.attrs["href"]
-        for tag in a_tags
-        if "href" in tag.attrs.keys() and tag.attrs["href"].endswith(".pdf")
-    ]
+    # Loop through the a-tags saving all href attributes that end with .pdf
+    hrefs = []
+    for tag in a_tags:
+        if "href" in tag.attrs.keys() and tag.attrs["href"].endswith(".pdf"):
+            hrefs.append(tag.attrs["href"])
 
+    # Convert all hrefs to URLs by adding the base URL if required
     return {href if href.startswith("http") else base_url + href for href in hrefs}
 
 
 @task
 def get_urls(url):
+    # Extract the base url: protocol + subdomain + domain
     base_url = get_base_url(url)
-    logger = get_run_logger()
 
+    # Stop processing if the request wasn't successful
     try:
         r = requests.get(url)
-        r.raise_for_status()
     except requests.RequestException as e:
-        logger.error(f"Error requesting URL {url}: {e}")
+        logger = get_run_logger()
+        logger.info(f"Error requesting URL... {url} - {e}")
+        return
+    if r.status_code != 200:
         return set()
 
-    bs = BeautifulSoup(r.content, "html.parser")
+    # Extract all a-tags from the website HTML
+    bs = BeautifulSoup(r.content)
     a_tags = bs.findAll("a")
 
+    # Extract content of href attribute from the a-tags and convert them into proper URLs using the base url
     hrefs = {tag.attrs["href"] for tag in a_tags if "href" in tag.attrs.keys()}
     hrefs = {href if href.startswith("http") else base_url + href for href in hrefs}
 
+    # Return only URLs that don't end with .pdf and that start with the base URL
     return {
         href
         for href in hrefs
@@ -96,53 +76,65 @@ def get_urls(url):
 
 
 @task
-def download_pdf(pdf_url, save_path):
-    logger = get_run_logger()
-    original_filename = os.path.basename(pdf_url).split("/")[-1]
-    rand_num = str(uuid.uuid4())
-    new_filename = original_filename[:-4] + "_" + rand_num + original_filename[-4:]
-    file = Path(new_filename)
-    try:
-        logger.info(f"Downloading PDF: {pdf_url}")
-        r = requests.get(pdf_url, stream=True)
-        r.raise_for_status()
-        put_data_to_s3(r.content, f"{save_path}/{new_filename}")
-        logger.info(f"Successfully uploaded PDF to S3: {file}")
-    except requests.RequestException as e:
-        logger.error(f"Error downloading PDF {pdf_url}: {e}")
+def download_pdfs(urls, save_path, print_details=False):
+    # Create/Save path where to store all PDFs
+    path = Path(save_path)
+    path.mkdir(parents=True, exist_ok=True)
 
+    # Loop through the set of PDF URLs, get their contents and save the files
+    for i, pdf_url in enumerate(urls):
 
-@task
-def download_pdfs(urls, save_path):
-    logger = get_run_logger()
-    logger.info(f"Starting download of {len(urls)} PDFs.")
+        # Extract the original file name from the PDF in the website
+        original_filename = os.path.basename(pdf_url).split("/")[-1]
 
-    # Submit all download tasks concurrently
-    download_tasks = [download_pdf.submit(url, save_path) for url in urls]
+        # Set up a new name by appending a random number to the original name (this avoids name duplication)
+        rand_num = str(random.randrange(100000, 199999))
+        new_filename = original_filename[:-4] + "_" + rand_num + original_filename[-4:]
+        file = Path(new_filename)
 
-    for download_task in download_tasks:
-        download_task.result()
+        # Start downloading the PDF, stop processing if the request wasn't successful
+        logger = get_run_logger()
+        if print_details:
+            logger.info(f"Downloading ({i+1}/{len(urls)}) PDF... ")
+        try:
+            r = requests.get(pdf_url, stream=True)
+        except requests.RequestException:
+            logger.info("Error downloading PDF... " + pdf_url)
+            continue
+        if r.status_code != 200:
+            continue
+
+        # If the file content was retrieved successfully, then write & save the new PDF file
+        with open(path.joinpath(file), "wb") as f:
+            f.write(r.content)
+        if print_details:
+            logger.info("Successful... " + new_filename)
 
 
 @flow
 def download_pdfs_from_url_recursive(
-    url, save_path, remaining_levels, original_levels, unique_pdfs
+    url, save_path, remaining_levels, original_levels, unique_pdfs, print_details=False
 ):
-    logger = get_run_logger()
+    # Get all PDF URLs and work only with the ones not previously found
     pdf_urls = get_pdf_urls(url)
     pdf_urls = [f for f in pdf_urls if f not in unique_pdfs]
     unique_pdfs.update(pdf_urls)
 
-    logger.info(
-        f"Depth Level 0 (Main Source) -> {len(pdf_urls)} PDFs found -> "
-        f"Source: {url}"
-    )
+    # Print PDFs found in the main source
+    logger = get_run_logger()
+    if remaining_levels == original_levels and print_details:
+        logger.info(
+            f"Depth Level 0 (Main Source) -> {len(pdf_urls)} PDFs found -> Source: {url}"
+        )
 
-    download_pdfs(pdf_urls, save_path)
+    # Download PDFs
+    download_pdfs(pdf_urls, save_path, print_details)
 
+    # If there's no remaining levels to dive, stop processing
     if remaining_levels == 0:
         return
 
+    # Loop through the other URLs extracting PDF URLs from each one of them
     remaining_levels -= 1
     other_urls = get_urls(url)
     for i, url_inside in enumerate(other_urls):
@@ -152,28 +144,55 @@ def download_pdfs_from_url_recursive(
             remaining_levels,
             original_levels,
             unique_pdfs,
+            print_details,
         )
         depth_level = original_levels - remaining_levels
-        logger.info(
-            f"Depth Level {depth_level} -> {i + 1}/{len(other_urls)} URLs -> "
-            f"Source: {url_inside}"
-        )
+        if print_details:
+            logger.info(
+                "..." * (depth_level - 1)
+                + f"Depth Level {depth_level} -> {i + 1}/{len(other_urls)} URLs -> Source: {url_inside}"
+            )
 
 
 @flow
-def download_pdfs_from_source(source_path: str, save_path: str, levels: int) -> None:
-    sources = fetch_urls_from_s3(file="data/source.txt")
-    for i, link in enumerate(sources):
-        logger = get_run_logger()
-        logger.info(f"Extracting Main Source #{i+1}: {link}")
+def download_pdfs_from_source_txt():
+    # TEMP PARAMS ######################
+    source_path = "data/source.txt"
+    save_path = "data"
+    levels = 0
+    print_details = True
+    #####################################
+
+    # Set proper a path for the source txt and the saving path
+    proper_source_path = Path(source_path)
+    proper_save_path = Path(save_path)
+
+    # Open the txt file that contains all URLs to explore
+    source_txt = open(proper_source_path, "r")
+
+    # Loop through the URLs, downloading the pdfs >=0 levels deep
+    for i, line in enumerate(source_txt):
+        link = line.strip()
+
+        # Print details if required
+        if print_details:
+            logger = get_run_logger()
+            logger.info(f"Extracting Main Source #{i+1}: {link}")
+
+        # Download all PDF files from each URL >=0 levels deep, into a specific folder for this URL only
+        source_save_path = Path(proper_save_path, str(i + 1))
         download_pdfs_from_url_recursive(
             link,
-            save_path,
+            source_save_path,
             remaining_levels=levels,
             original_levels=levels,
             unique_pdfs=set(),
+            print_details=print_details,
         )
+
+    # Close the txt file
+    source_txt.close()
 
 
 if __name__ == "__main__":
-    download_pdfs_from_source(source_path="data", save_path="data/raw", levels=0)
+    pass
